Fixed augmentation when changing canonical visit
This fixes the way that we update augmentation data when changing EulerTourVertex.arbitraryVisit. Previously, we were calling augmentAncestorFlags() to update the augmentation data. However, changing a vertex's canonical visit affects both the hasGraphEdge and hasForestEdge flags and the user's augmentation data. We have to call augment() rather than augmentFlags() on the ancestors.