Remove node before changing its vertex's canonical visit
This changes addForestEdge to remove the maximum node before changing its vertex's canonical visit.

Changing a vertex's canonical visit can affect a node's augmentation data. Normally, we have to call augment() on a node's ancestors when its augmentation data change, but since we are about to remove the node, we don't. I'm pretty sure that the behavior is correct as written - where we change the node's augmentation data and then remove it without explicitly calling augment() on its ancestors. However, it's much clearer that the behavior is correct if we remove the node first. It's also more elegant to remove the node first, because we don't temporarily enter a state where the nodes' augmentation data are potentially broken.

Furthermore, the way it's currently written, it's possible that a single augmentation value will be combined into a given (temporary) result twice. This shouldn't be a problem in principle, but for certain types of combining functions, it might be unexpected or confusing. For example, if we are augmenting a game map with the strongest monster in a connected component, it might be unusual for the same exact monster M to be passed in for both arguments of Augmentation.combine (i.e. combine(M, M)). In extreme cases, this violation of expectations could result in "combine" throwing an exception, if it wasn't written to account for this possibility.